First commit: alpha version of the SOI API client
This commit includes the base skeletion for the client gem for
the Single Offender Identity API ruby client.

The base functionality included in this alpha version is:
- client initialization with mandatory configurations (presently,
client_id, client_secret and base_url)
- Authorization request to retrieve the access token for further
authenticated requests. The access token is cached in memory and
re-used while is still valid. Once expired, a new authorization
request is performed to re-new the access token.
- GET request, which leverages from the authorization request to
guarantee that all requests have an access token before they're
executed. Currently, it only supports GET params as arguments.
- Being an alpha version, other request types (e.g. POST, PUT,
PATCH and DELETE are not yet supported)